rename dir 'common' -> 'arch'
