+++ coreblocks/frontend/decoder/rvc.py
@@ -4,7 +4,7 @@ from transactron import TModule
 from transactron.utils import ValueLike
 
 from coreblocks.params import *
-from coreblocks.common import *
+from coreblocks.arch import *
 
 
 # An instruction or an instruction with the valid signal
