+++ coreblocks/func_blocks/fu/common/rs.py
@@ -4,7 +4,7 @@ from amaranth import *
 from amaranth.lib.coding import PriorityEncoder
 from transactron import Method, Transaction, def_method, TModule
 from coreblocks.params import GenParams
-from coreblocks.common import OpType
+from coreblocks.arch import OpType
 from coreblocks.interface.layouts import RSLayouts
 from transactron.lib.metrics import HwExpHistogram, TaggedLatencyMeasurer
 from transactron.utils import RecordDict
