+++ coreblocks/func_blocks/fu/mul_unit.py
@@ -8,7 +8,7 @@ from coreblocks.func_blocks.fu.unsigned_multiplication.fast_recursive import Rec
 from coreblocks.func_blocks.fu.unsigned_multiplication.sequence import SequentialUnsignedMul
 from coreblocks.func_blocks.fu.unsigned_multiplication.shift import ShiftUnsignedMul
 from coreblocks.params import GenParams, FunctionalComponentParams
-from coreblocks.common import OpType, Funct3
+from coreblocks.arch import OpType, Funct3
 from coreblocks.interface.layouts import FuncUnitLayouts
 from transactron import *
 from transactron.core import def_method
