+++ coreblocks/interface/layouts.py
@@ -1,7 +1,7 @@
 from typing import Optional
 from amaranth.lib.data import StructLayout
 from coreblocks.params import GenParams
-from coreblocks.common import *
+from coreblocks.arch import *
 from transactron.utils import LayoutList, LayoutListField, layout_subset
 from transactron.utils.transactron_helpers import from_method_layout, make_layout
