+++ coreblocks/params/configurations.py
@@ -6,7 +6,7 @@ from dataclasses import dataclass, field
 from typing import Self
 from transactron.utils._typing import type_self_kwargs_as
 
-from coreblocks.common.isa import Extension
+from coreblocks.arch.isa import Extension
 from coreblocks.params.fu_params import BlockComponentParams
 
 from coreblocks.func_blocks.fu.common.rs_func_block import RSBlockComponent
