+++ coreblocks/priv/traps/exception.py
@@ -2,7 +2,7 @@ from amaranth import *
 from transactron.utils.dependencies import DependencyManager
 from coreblocks.params.genparams import GenParams
 
-from coreblocks.common import ExceptionCause
+from coreblocks.arch import ExceptionCause
 from coreblocks.interface.layouts import ExceptionRegisterLayouts
 from coreblocks.interface.keys import ExceptionReportKey
 from transactron.core import TModule, def_method, Method
