+++ test/frontend/test_instr_decoder.py
@@ -5,7 +5,7 @@ from transactron.testing import TestCaseWithSimulator
 from coreblocks.params import *
 from coreblocks.params.configurations import test_core_config
 from coreblocks.frontend.decoder.instr_decoder import InstrDecoder, Encoding, instructions_by_optype
-from coreblocks.common import *
+from coreblocks.arch import *
 from unittest import TestCase
 from typing import Optional
