+++ test/func_blocks/fu/test_div_unit.py
@@ -1,6 +1,6 @@
 from parameterized import parameterized_class
 
-from coreblocks.common import Funct3, Funct7, OpType
+from coreblocks.arch import Funct3, Funct7, OpType
 from coreblocks.func_blocks.fu.div_unit import DivFn, DivComponent
 
 from test.func_blocks.fu.functional_common import ExecFn, FunctionalUnitTestCase
