+++ test/func_blocks/fu/test_zbc.py
@@ -1,7 +1,7 @@
 from parameterized import parameterized_class
 
 from coreblocks.func_blocks.fu.zbc import ZbcFn, ZbcComponent
-from coreblocks.common import Funct3, Funct7, OpType
+from coreblocks.arch import Funct3, Funct7, OpType
 from coreblocks.params.configurations import test_core_config
 
 from test.func_blocks.fu.functional_common import ExecFn, FunctionalUnitTestCase
