+++ test/func_blocks/fu/test_zbs.py
@@ -1,4 +1,4 @@
-from coreblocks.common import Funct3, Funct7, OpType
+from coreblocks.arch import Funct3, Funct7, OpType
 from coreblocks.func_blocks.fu.zbs import ZbsFunction, ZbsComponent
 
 from test.func_blocks.fu.functional_common import ExecFn, FunctionalUnitTestCase
