+++ test/func_blocks/lsu/test_dummylsu.py
@@ -8,7 +8,7 @@ from transactron.utils import int_to_signed, signed_to_int
 from coreblocks.params import GenParams
 from coreblocks.func_blocks.fu.lsu.dummyLsu import LSUDummy
 from coreblocks.params.configurations import test_core_config
-from coreblocks.common import *
+from coreblocks.arch import *
 from coreblocks.interface.keys import ExceptionReportKey, InstructionPrecommitKey
 from transactron.utils.dependencies import DependencyManager
 from coreblocks.interface.layouts import ExceptionRegisterLayouts, RetirementLayouts
