+++ test/scheduler/test_rs_selection.py
@@ -6,8 +6,8 @@ from amaranth.sim import Settle, Passive
 
 from coreblocks.params import GenParams
 from coreblocks.interface.layouts import RSLayouts, SchedulerLayouts
-from coreblocks.common import Funct3, Funct7
-from coreblocks.common import OpType
+from coreblocks.arch import Funct3, Funct7
+from coreblocks.arch import OpType
 from coreblocks.params.configurations import test_core_config
 from coreblocks.scheduler.scheduler import RSSelection
 from transactron.lib import FIFO, Adapter, AdapterTrans
